mcp_agent/requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
mcp-server==0.5.0
python-dotenv==1.0.1
aiocache==0.12.2
